+++ front/src/utils/editorjs/utils.js
@@ -1,4 +1,4 @@
-import request from 'axios';
+import api from '../api';
 
 export const createInput = ({
   readOnly = false,
@@ -45,34 +45,25 @@ export const sanitizeBlocks = {
   mark: true,
 };
 
-export class Uploader {
-  constructor({ config, onSuccess, onError }) {
-    this.config = config;
-    this.onSuccess = onSuccess;
-    this.onError = onError;
-  }
+export const uploadFile = async ({ parent, config, onSuccess, onError }) => {
+  try {
+    const formData = new FormData();
+    const {
+      files: [file],
+    } = parent;
+    formData.append('file', file);
 
-  async uploadFile(parent) {
-    try {
-      const formData = new FormData();
-      const {
-        files: [file],
-      } = parent;
-      formData.append('file', file);
-
-      const response = await request.post(
-        `${process.env.REACT_APP_SB_HOST}/api/v1/files`,
-        formData,
-        {
-          headers: {
-            ...this.config.headers,
-            'content-type': 'multipart/form-data',
-          },
+    const response = await api.post(
+      `${process.env.REACT_APP_SB_HOST}/api/v1/files`,
+      formData,
+      {
+        headers: {
+          'content-type': 'multipart/form-data',
         },
-      );
-      this.onSuccess(response);
-    } catch (e) {
-      this.onError(e);
-    }
+      },
+    );
+    onSuccess(response);
+  } catch (e) {
+    onError(e);
   }
-}
+};
